+++ CHANGELOG.md
@@ -8,6 +8,7 @@ omnibus-gitlab repository.
 - Switch the redis user's shell to /bin/false
 - NGINX listen on IPv6 by default (George Gooden)
 - Update Redis to 3.2.5 (Takuya Noguchi)
+- Updarted cacerts.pem to 2016-11-02 version
 
 8.13.4